backend/requirements.txt
# MeetAI Backend — Python 3.11+
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
greenlet==3.3.2
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
httpx==0.28.1
groq==0.13.1

# Local Whisper (pulls torch; use tiny/base for lighter installs)
openai-whisper>=20231117
# Bundled ffmpeg binary so Whisper can decode MP3/WAV on Windows without a system install
imageio-ffmpeg>=0.5.1

# Structured deadlines / reminders
dateparser==1.2.0
apscheduler==3.11.0

# Notifications (SMTP). Gracefully no-ops when SMTP_HOST is unset.
aiosmtplib==3.0.2

# Background worker. In-process fallback when ARQ_REDIS_URL is unset.
arq==0.26.3

# Vector search on PostgreSQL. Falls back to a cosine Python path on SQLite.
pgvector==0.3.6

# Optional better embeddings — install separately to opt in:
#   pip install sentence-transformers
# MeetAI auto-detects and uses it when available; otherwise a deterministic
# hash-token fallback is used so the app still works offline.
